requirements.txt
absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Bottleneck
cachetools==5.2.0
certifi
charset-normalizer==2.1.1
contourpy==1.0.6
cycler
et-xmlfile==1.1.0
filelock==3.9.0
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
hugsvision==0.75.3
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing
kiwisolver
libclang==14.0.6
lightning-utilities==0.5.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multidict==6.0.4
munkres==1.1.4
numexpr
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.68
openpyxl==3.0.10
opt-einsum==3.3.0
packaging
pandas==1.5.2
Pillow==9.4.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pyparsing
python-dateutil
pytorch-lightning==1.8.6
pytz
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn
six
split-folders==0.5.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.13.2
torch==1.9.1+cu111
torchaudio==0.9.1
torchmetrics==0.11.0
torchvision==0.10.1+cu111
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.6
yarl==1.8.2
zipp==3.11.0
